\name{LearnRcpp2-package}
\alias{LearnRcpp2-package}
\alias{LearnRcpp2}
\docType{package}
\title{\packageTitle{LearnRcpp2}}
\description{\packageDescription{LearnRcpp2}}
\section{Package Content}{\packageIndices{LearnRcpp2}}
\author{\packageAuthor{LearnRcpp2}}
\section{Maintainer}{\packageMaintainer{LearnRcpp2}}
\keyword{package}
